revision 7 beta
This has some state transitions and a basic game laid out.
Still needs levels and loading better handled.

+++ trunk/src/com/kobaj/runfoxrun/FPSManager.java
@@ -37,8 +37,8 @@ public class FPSManager
 		return fps;
 	}
 	
-	public int getDelta()
+	public float getDelta()
 	{
-		return (int) Math.max(Math.min(delta, 10000), 0);
+		return Math.max(Math.min(delta, 10000), 0);
 	}
 }
